# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HSE Mobile Application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plants:
    get:
      operationId: getPlants
      tags: [plants]
      summary: List all plants
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of plants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plant"
    post:
      operationId: createPlant
      tags: [plants]
      summary: Create a plant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlantRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
  /plants/{id}:
    put:
      operationId: updatePlant
      tags: [plants]
      summary: Update plant
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlantRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
    delete:
      operationId: deletePlant
      tags: [plants]
      summary: Delete plant
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /actions:
    get:
      operationId: getActions
      tags: [actions]
      summary: List master actions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of actions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Action"
    post:
      operationId: createAction
      tags: [actions]
      summary: Create action
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Action"
  /actions/{id}:
    put:
      operationId: updateAction
      tags: [actions]
      summary: Update action
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActionRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Action"
    delete:
      operationId: deleteAction
      tags: [actions]
      summary: Delete action
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /incidents:
    get:
      operationId: getIncidents
      tags: [incidents]
      summary: List incidents
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: plantId
          in: query
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Create incident report
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncidentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{id}:
    get:
      operationId: getIncidentById
      tags: [incidents]
      summary: Get incident
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
    put:
      operationId: updateIncident
      tags: [incidents]
      summary: Update incident status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIncidentRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Plant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        description:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - code
        - createdAt
    CreatePlantRequest:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
      required:
        - name
        - code
    Action:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    CreateActionRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
...
    Incident:
      type: object
      properties:
        id:
          type: integer
        reporterId:
          type: integer
        reporterName:
          type: string
        plantId:
          type: integer
        plantName:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryRiskLevel:
          type: string
          enum: [low, medium, high]
        incidentDate:
          type: string
        reportedDate:
          type: string
        detail:
          type: string
        actionId:
          type: integer
        actionName:
          type: string
        needsFurtherAction:
          type: boolean
        status:
          type: string
          enum: [open, in_progress, closed]
        closedAt:
          type: string
        picGroupId:
          type: integer
        picGroupName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - reporterId
        - reporterName
        - plantId
        - plantName
        - categoryId
        - categoryName
        - categoryRiskLevel
        - incidentDate
        - reportedDate
        - detail
        - needsFurtherAction
        - status
        - createdAt
    CreateIncidentRequest:
      type: object
      properties:
        reporterId:
          type: integer
        plantId:
          type: integer
        categoryId:
          type: integer
        incidentDate:
          type: string
        detail:
          type: string
        actionId:
          type: integer
        needsFurtherAction:
          type: boolean
      required:
        - reporterId
        - plantId
        - categoryId
        - incidentDate
        - detail
        - needsFurtherAction
    UpdateIncidentRequest:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, closed]
        actionId:
          type: integer
        needsFurtherAction:
          type: boolean
        detail:
          type: string
...